feat(db): 添加Flyway数据库迁移配置

+++ src/main/kotlin/top/inept/blog/feature/comment/model/vo/CommentReplyVO.kt
@@ -12,7 +12,7 @@ data class CommentReplyVO(
     val content: String,
 
     @field:Schema(description = "openapi.comment.like_count")
-    val likeCount: Int,
+    val likeCount: Long,
 
     @field:Schema(description = "openapi.comment.user")
     val user: UserLiteVO,
